+++ conjured_item_processor.rb
@@ -1,0 +1,12 @@
+require 'item_processor'
+
+# ConjuredItemProcessor
+#
+# Updates a conjured item
+class ConjuredItemProcessor < ItemProcessor
+  def update
+    item.sell_in -= 1
+    item.quality -= 2 unless item.quality.zero?
+    item
+  end
+end
