+++ item_processor.rb
@@ -1,0 +1,13 @@
+# ItemProcessor
+#
+# Abstract class to update an item
+class ItemProcessor
+  attr_reader :item
+
+  def initialize(item)
+    @item = item
+  end
+
+  def update
+  end
+end
