+++ sulfuras_item_processor.rb
@@ -1,0 +1,7 @@
+require 'item_processor'
+
+# SulfurasItemProcessor
+#
+# Updates sulfuras
+class SulfurasItemProcessor < ItemProcessor
+end
